Upload annotations: Add tests for CPC image-name matching. Also split up the tests into more files.
