+++ project/upload/tests/test_annotations.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
+#
+# Tests that generally apply to both CSV and CPC annotation uploads.
+
 from __future__ import unicode_literals
 import codecs
-from io import StringIO
 
 from django.core.files.base import ContentFile
 from django.urls import reverse
@@ -11,9 +13,8 @@ from annotations.model_utils import AnnotationAreaUtils
 from annotations.models import Annotation
 from images.model_utils import PointGen
 from images.models import Point
-from lib.tests.utils import (
-    BasePermissionTest, ClientTest, sample_image_as_file)
-from .utils import UploadAnnotationsTestMixin
+from lib.tests.utils import BasePermissionTest, ClientTest
+from .utils import UploadAnnotationsBaseTest
 
 
 class PermissionTest(BasePermissionTest):
@@ -139,10 +140,6 @@ class UploadAnnotationsNoLabelsetTest(ClientTest):
             response,
             "You must create a labelset before uploading annotations.")
         self.assertTemplateUsed(response, 'labels/labelset_required.html')
-
-
-class UploadAnnotationsBaseTest(ClientTest, UploadAnnotationsTestMixin):
-    pass
 
 
 class UploadAnnotationsTest(UploadAnnotationsBaseTest):
@@ -1510,615 +1507,3 @@ class UploadAnnotationsFormatTest(UploadAnnotationsBaseTest):
             preview_response.json(),
             dict(error="The submitted file is empty."),
         )
-
-
-class UploadAnnotationsCSVFormatTest(UploadAnnotationsBaseTest):
-    """
-    Tests (mostly error cases) specific to CSV format.
-    """
-    @classmethod
-    def setUpTestData(cls):
-        super(UploadAnnotationsCSVFormatTest, cls).setUpTestData()
-
-        cls.user = cls.create_user()
-        cls.source = cls.create_source(cls.user)
-        labels = cls.create_labels(cls.user, ['A', 'B'], 'Group1')
-        cls.create_labelset(cls.user, cls.source, labels)
-
-        cls.img1 = cls.upload_image(
-            cls.user, cls.source,
-            image_options=dict(filename='1.png', width=100, height=100))
-
-    def check(self, preview_response, upload_response):
-
-        self.assertDictEqual(
-            preview_response.json(),
-            dict(
-                success=True,
-                previewTable=[
-                    dict(
-                        name=self.img1.metadata.name,
-                        link=reverse(
-                            'annotation_tool',
-                            kwargs=dict(image_id=self.img1.pk)),
-                        createInfo="Will create 1 points, 1 annotations",
-                    ),
-                ],
-                previewDetails=dict(
-                    numImages=1,
-                    totalPoints=1,
-                    totalAnnotations=1,
-                    numImagesWithExistingAnnotations=0,
-                ),
-            ),
-        )
-        self.assertDictEqual(upload_response.json(), dict(success=True))
-
-        values_set = set(
-            Point.objects.filter(image__in=[self.img1])
-            .values_list('column', 'row', 'point_number', 'image_id'))
-        self.assertSetEqual(values_set, {
-            (60, 40, 1, self.img1.pk),
-        })
-
-        annotations = Annotation.objects.filter(image__in=[self.img1])
-        values_set = set(
-            (a.label_code, a.point.pk, a.image.pk)
-            for a in annotations
-        )
-        self.assertSetEqual(values_set, {
-            ('A', Point.objects.get(
-                point_number=1, image=self.img1).pk, self.img1.pk),
-        })
-
-    def test_skipped_csv_columns(self):
-        """
-        The CSV can have column names that we don't recognize. Those columns
-        will just be ignored.
-        """
-        rows = [
-            ['Name', 'Column', 'Row', 'Annotator', 'Label'],
-            ['1.png', 60, 40, 'Jane', 'A'],
-        ]
-        csv_file = self.make_csv_file('A.csv', rows)
-        preview_response = self.preview_csv_annotations(
-            self.user, self.source, csv_file)
-        upload_response = self.upload_annotations(self.user, self.source)
-
-        self.check(preview_response, upload_response)
-
-    def test_columns_different_order(self):
-        """
-        The CSV columns can be in a different order.
-        """
-        rows = [
-            ['Row', 'Name', 'Label', 'Column'],
-            [40, '1.png', 'A', 60],
-        ]
-        csv_file = self.make_csv_file('A.csv', rows)
-        preview_response = self.preview_csv_annotations(
-            self.user, self.source, csv_file)
-        upload_response = self.upload_annotations(self.user, self.source)
-
-        self.check(preview_response, upload_response)
-
-    def test_columns_different_case(self):
-        """
-        The CSV column names can use different upper/lower case and still
-        be matched to the expected column names.
-        """
-        rows = [
-            ['name', 'coLUmN', 'ROW', 'Label'],
-            ['1.png', 60, 40, 'A'],
-        ]
-        csv_file = self.make_csv_file('A.csv', rows)
-        preview_response = self.preview_csv_annotations(
-            self.user, self.source, csv_file)
-        upload_response = self.upload_annotations(self.user, self.source)
-
-        self.check(preview_response, upload_response)
-
-    def test_no_name_column(self):
-        """
-        No CSV columns correspond to the name field.
-        """
-        rows = [
-            ['Column', 'Row', 'Label'],
-            [50, 50, 'A'],
-            [60, 40, 'B'],
-        ]
-        csv_file = self.make_csv_file('A.csv', rows)
-        preview_response = self.preview_csv_annotations(
-            self.user, self.source, csv_file)
-
-        self.assertDictEqual(
-            preview_response.json(),
-            dict(error="CSV must have a column called Name"),
-        )
-
-    def test_no_row_column(self):
-        """
-        No CSV columns correspond to the row field.
-        """
-        rows = [
-            ['Name', 'Column', 'Label'],
-            ['1.png', 50, 'A'],
-            ['1.png', 60, 'B'],
-        ]
-        csv_file = self.make_csv_file('A.csv', rows)
-        preview_response = self.preview_csv_annotations(
-            self.user, self.source, csv_file)
-
-        self.assertDictEqual(
-            preview_response.json(),
-            dict(error="CSV must have a column called Row"),
-        )
-
-    def test_no_column_column(self):
-        """
-        No CSV columns correspond to the column field.
-        """
-        rows = [
-            ['Name', 'Row'],
-            ['1.png', 50],
-            ['1.png', 40],
-        ]
-        csv_file = self.make_csv_file('A.csv', rows)
-        preview_response = self.preview_csv_annotations(
-            self.user, self.source, csv_file)
-
-        self.assertDictEqual(
-            preview_response.json(),
-            dict(error="CSV must have a column called Column"),
-        )
-
-    def test_missing_row(self):
-        """
-        A row is missing the row field.
-        """
-        rows = [
-            ['Name', 'Column', 'Row'],
-            ['1.png', 50, 50],
-            ['1.png', 60, ''],
-        ]
-        csv_file = self.make_csv_file('A.csv', rows)
-        preview_response = self.preview_csv_annotations(
-            self.user, self.source, csv_file)
-
-        self.assertDictEqual(
-            preview_response.json(),
-            dict(error="CSV row 3 is missing a Row value"),
-        )
-
-    def test_missing_column(self):
-        """
-        A row is missing the column field.
-        """
-        rows = [
-            ['Name', 'Column', 'Row'],
-            ['1.png', '', 50],
-            ['1.png', 60, 40],
-        ]
-        csv_file = self.make_csv_file('A.csv', rows)
-        preview_response = self.preview_csv_annotations(
-            self.user, self.source, csv_file)
-
-        self.assertDictEqual(
-            preview_response.json(),
-            dict(error="CSV row 2 is missing a Column value"),
-        )
-
-    def test_field_with_newline(self):
-        """Field value with a newline character in it (within quotation marks).
-        There's no reason to have this in any of the recognized columns, so we
-        use an extra ignored column to test."""
-        content = (
-            'Name,Column,Row,Comments,Label'
-            '\n1.png,60,40,"Here are\nsome comments.",A'
-        )
-        csv_file = ContentFile(content, name='A.csv')
-        preview_response = self.preview_csv_annotations(
-            self.user, self.source, csv_file)
-        upload_response = self.upload_annotations(self.user, self.source)
-
-        self.check(preview_response, upload_response)
-
-    def test_field_with_surrounding_quotes(self):
-        content = (
-            'Name,Column,Row,Label'
-            '\n"1.png","60","40","A"'
-        )
-        csv_file = ContentFile(content, name='A.csv')
-        preview_response = self.preview_csv_annotations(
-            self.user, self.source, csv_file)
-        upload_response = self.upload_annotations(self.user, self.source)
-
-        self.check(preview_response, upload_response)
-
-    def test_field_with_surrounding_whitespace(self):
-        """Strip whitespace surrounding the CSV values."""
-        content = (
-            'Name ,\tColumn\t,  Row,\tLabel    '
-            '\n\t1.png,    60   , 40,A'
-        )
-        csv_file = ContentFile(content, name='A.csv')
-        preview_response = self.preview_csv_annotations(
-            self.user, self.source, csv_file)
-        upload_response = self.upload_annotations(self.user, self.source)
-
-        self.check(preview_response, upload_response)
-
-    def test_non_csv(self):
-        """
-        Do at least basic detection of non-CSV files.
-        """
-        f = sample_image_as_file('A.jpg')
-        preview_response = self.preview_csv_annotations(
-            self.user, self.source, f)
-
-        self.assertDictEqual(
-            preview_response.json(),
-            dict(
-                error="The selected file is not a CSV file.",
-            ),
-        )
-
-
-class UploadAnnotationsCPCFormatTest(UploadAnnotationsBaseTest):
-    """
-    Tests (mostly error cases) specific to CPC format.
-    """
-    @classmethod
-    def setUpTestData(cls):
-        super(UploadAnnotationsCPCFormatTest, cls).setUpTestData()
-
-        cls.user = cls.create_user()
-        cls.source = cls.create_source(cls.user)
-        labels = cls.create_labels(cls.user, ['A', 'B'], 'Group1')
-        cls.create_labelset(cls.user, cls.source, labels)
-
-        cls.img1 = cls.upload_image(
-            cls.user, cls.source,
-            image_options=dict(filename='1.png', width=100, height=100))
-
-    def test_one_line(self):
-        stream = StringIO()
-        # Line 1
-        stream.writelines(['a,b,c,d,e,f\n'])
-        cpc_file = ContentFile(stream.getvalue(), name='1.cpc')
-        preview_response = self.preview_cpc_annotations(
-            self.user, self.source, [cpc_file])
-
-        self.assertDictEqual(
-            preview_response.json(),
-            dict(error="File 1.cpc seems to have too few lines."))
-
-    def test_line_1_not_enough_tokens(self):
-        stream = StringIO()
-        # Line 1
-        stream.writelines(['abc\n'])
-        cpc_file = ContentFile(stream.getvalue(), name='1.cpc')
-        preview_response = self.preview_cpc_annotations(
-            self.user, self.source, [cpc_file])
-
-        self.assertDictEqual(
-            preview_response.json(),
-            dict(error=(
-                "File 1.cpc, line 1 has"
-                " 1 comma-separated tokens, but"
-                " 6 were expected.")))
-
-    def test_line_1_too_many_tokens(self):
-        stream = StringIO()
-        # Line 1
-        stream.writelines(['ab,cd,ef,gh,ij,kl,mn\n'])
-        cpc_file = ContentFile(stream.getvalue(), name='1.cpc')
-        preview_response = self.preview_cpc_annotations(
-            self.user, self.source, [cpc_file])
-
-        self.assertDictEqual(
-            preview_response.json(),
-            dict(error=(
-                "File 1.cpc, line 1 has"
-                " 7 comma-separated tokens, but"
-                " 6 were expected.")))
-
-    def test_line_1_quoted_commas_accepted(self):
-        """Any commas between quotes should be considered part of a
-        token, instead of a token separator."""
-        stream = StringIO()
-        # Line 1
-        stream.writelines(['ab,cd,ef,"gh,ij",kl,mn\n'])
-        cpc_file = ContentFile(stream.getvalue(), name='1.cpc')
-        preview_response = self.preview_cpc_annotations(
-            self.user, self.source, [cpc_file])
-
-        # Should get past line 1 just fine, and then error due to not
-        # having more lines.
-        self.assertDictEqual(
-            preview_response.json(),
-            dict(error="File 1.cpc seems to have too few lines."))
-
-    def test_line_2_wrong_number_of_tokens(self):
-        stream = StringIO()
-        # Line 1
-        stream.writelines(['a,b,c,d,e,f\n'])
-        # Line 2
-        stream.writelines(['1,2,3\n'])
-        cpc_file = ContentFile(stream.getvalue(), name='1.cpc')
-        preview_response = self.preview_cpc_annotations(
-            self.user, self.source, [cpc_file])
-
-        self.assertDictEqual(
-            preview_response.json(),
-            dict(error=(
-                "File 1.cpc, line 2 has"
-                " 3 comma-separated tokens, but"
-                " 2 were expected.")))
-
-    def test_line_6_wrong_number_of_tokens(self):
-        stream = StringIO()
-        # Line 1
-        stream.writelines(['a,b,c,d,e,f\n'])
-        # Lines 2-5
-        stream.writelines(['1,2\n']*4)
-        # Line 6
-        stream.writelines(['abc,def\n'])
-        cpc_file = ContentFile(stream.getvalue(), name='1.cpc')
-        preview_response = self.preview_cpc_annotations(
-            self.user, self.source, [cpc_file])
-
-        self.assertDictEqual(
-            preview_response.json(),
-            dict(error=(
-                "File 1.cpc, line 6 has"
-                " 2 comma-separated tokens, but"
-                " 1 were expected.")))
-
-    def test_line_6_not_number(self):
-        stream = StringIO()
-        # Line 1
-        stream.writelines(['a,b,c,d,e,f\n'])
-        # Lines 2-5
-        stream.writelines(['1,2\n']*4)
-        # Line 6
-        stream.writelines(['abc\n'])
-        cpc_file = ContentFile(stream.getvalue(), name='1.cpc')
-        preview_response = self.preview_cpc_annotations(
-            self.user, self.source, [cpc_file])
-
-        self.assertDictEqual(
-            preview_response.json(),
-            dict(error=(
-                "File 1.cpc, line 6 is supposed to have"
-                " the number of points, but this line isn't a"
-                " positive integer: abc")))
-
-    def test_line_6_number_below_1(self):
-        stream = StringIO()
-        # Line 1
-        stream.writelines(['a,b,c,d,e,f\n'])
-        # Lines 2-5
-        stream.writelines(['1,2\n']*4)
-        # Line 6
-        stream.writelines(['0\n'])
-        cpc_file = ContentFile(stream.getvalue(), name='1.cpc')
-        preview_response = self.preview_cpc_annotations(
-            self.user, self.source, [cpc_file])
-
-        self.assertDictEqual(
-            preview_response.json(),
-            dict(error=(
-                "File 1.cpc, line 6 is supposed to have"
-                " the number of points, but this line isn't a"
-                " positive integer: 0")))
-
-    def test_point_position_line_wrong_number_of_tokens(self):
-        stream = StringIO()
-        # Line 1
-        stream.writelines(['a,b,c,d,e,f\n'])
-        # Lines 2-5
-        stream.writelines(['1,2\n']*4)
-        # Line 6
-        stream.writelines(['10\n'])
-        # Line 7-15: point positions (one line too few)
-        stream.writelines([
-            '{n},{n}\n'.format(n=n*15) for n in range(1, 9+1)])
-        # Line 16: labels
-        stream.writelines(['a,b,c,d\n'])
-        cpc_file = ContentFile(stream.getvalue(), name='1.cpc')
-        preview_response = self.preview_cpc_annotations(
-            self.user, self.source, [cpc_file])
-
-        self.assertDictEqual(
-            preview_response.json(),
-            dict(error=(
-                "File 1.cpc, line 16 has"
-                " 4 comma-separated tokens, but"
-                " 2 were expected.")))
-
-    def test_label_line_wrong_number_of_tokens(self):
-        stream = StringIO()
-        # Line 1
-        stream.writelines(['a,b,c,d,e,f\n'])
-        # Lines 2-5
-        stream.writelines(['1,2\n']*4)
-        # Line 6
-        stream.writelines(['10\n'])
-        # Lines 7-17: point positions (one line too many)
-        stream.writelines([
-            '{n},{n}\n'.format(n=n*15) for n in range(1, 11+1)])
-        cpc_file = ContentFile(stream.getvalue(), name='1.cpc')
-        preview_response = self.preview_cpc_annotations(
-            self.user, self.source, [cpc_file])
-
-        self.assertDictEqual(
-            preview_response.json(),
-            dict(error=(
-                "File 1.cpc, line 17 has"
-                " 2 comma-separated tokens, but"
-                " 4 were expected.")))
-
-    def test_not_enough_label_lines(self):
-        stream = StringIO()
-        # Line 1
-        stream.writelines(['a,b,c,d,e,f\n'])
-        # Lines 2-5
-        stream.writelines(['1,2\n']*4)
-        # Line 6
-        stream.writelines(['10\n'])
-        # Lines 7-16: point positions
-        stream.writelines([
-            '{n},{n}\n'.format(n=n*15) for n in range(1, 10+1)])
-        # Line 17-25: labels (one line too few)
-        stream.writelines(['a,b,c,d\n']*9)
-        cpc_file = ContentFile(stream.getvalue(), name='1.cpc')
-        preview_response = self.preview_cpc_annotations(
-            self.user, self.source, [cpc_file])
-
-        self.assertDictEqual(
-            preview_response.json(),
-            dict(error="File 1.cpc seems to have too few lines."))
-
-    def test_multiple_cpcs_for_one_image(self):
-        stream = StringIO()
-        # Line 1
-        stream.writelines(['a,"D:\\Panama transects\\1.png",c,d,e,f\n'])
-        # Lines 2-5
-        stream.writelines(['1,2\n']*4)
-        # Line 6
-        stream.writelines(['10\n'])
-        # Lines 7-16: point positions
-        # Multiply by 15 so they end up on different pixels. Otherwise
-        # we may get the 'multiple points on same pixel' error instead.
-        stream.writelines([
-            '{n},{n}\n'.format(n=n*15) for n in range(1, 10+1)])
-        # Line 17-26: labels
-        stream.writelines(['a,b,c,d\n']*10)
-        cpc_file_1 = ContentFile(stream.getvalue(), name='1.cpc')
-
-        stream = StringIO()
-        # Line 1
-        stream.writelines(['a,"D:\\GBR transects\\1.png",c,d,e,f\n'])
-        # Lines 2-5
-        stream.writelines(['1,2\n']*4)
-        # Line 6
-        stream.writelines(['10\n'])
-        # Lines 7-16: point positions
-        stream.writelines([
-            '{n},{n}\n'.format(n=n*15) for n in range(1, 10+1)])
-        # Line 17-26: labels
-        stream.writelines(['a,b,c,d\n']*10)
-        cpc_file_2 = ContentFile(stream.getvalue(), name='2.cpc')
-
-        preview_response = self.preview_cpc_annotations(
-            self.user, self.source, [cpc_file_1, cpc_file_2])
-
-        self.assertDictEqual(
-            preview_response.json(),
-            dict(error=(
-                "Image 1.png has points from more than one .cpc file: 1.cpc"
-                " and 2.cpc. There should be only one .cpc file"
-                " per image.")))
-
-
-class UploadAnnotationsSaveCPCInfoTest(UploadAnnotationsBaseTest):
-    """
-    Tests saving of CPC file info when uploading CPCs.
-    """
-    @classmethod
-    def setUpTestData(cls):
-        super(UploadAnnotationsSaveCPCInfoTest, cls).setUpTestData()
-
-        cls.user = cls.create_user()
-        cls.source = cls.create_source(cls.user)
-        labels = cls.create_labels(cls.user, ['A', 'B'], 'Group1')
-        cls.create_labelset(cls.user, cls.source, labels)
-
-        cls.img1 = cls.upload_image(
-            cls.user, cls.source,
-            image_options=dict(filename='1.jpg', width=100, height=100))
-        cls.img2 = cls.upload_image(
-            cls.user, cls.source,
-            image_options=dict(filename='2.jpg', width=100, height=100))
-
-    def test_cpc_content_one_image(self):
-        cpc_files = [
-            self.make_cpc_file(
-                '1.cpc',
-                r"C:\My Photos\2017-05-13 GBR\1.jpg", [
-                    (49*15, 49*15, 'A'),
-                    (59*15, 39*15, 'B')]),
-        ]
-        # Save the file content for comparison purposes.
-        img1_expected_cpc_content = cpc_files[0].read()
-        # Reset the file pointer so that the views can read from the start.
-        cpc_files[0].seek(0)
-
-        self.preview_cpc_annotations(self.user, self.source, cpc_files)
-        self.upload_annotations(self.user, self.source)
-
-        self.img1.refresh_from_db()
-        self.assertEqual(self.img1.cpc_content, img1_expected_cpc_content)
-        self.assertEqual(self.img1.cpc_filename, '1.cpc')
-        self.img2.refresh_from_db()
-        self.assertEqual(self.img2.cpc_content, '')
-        self.assertEqual(self.img2.cpc_filename, '')
-
-    def test_cpc_content_multiple_images(self):
-        cpc_files = [
-            self.make_cpc_file(
-                'GBR_1.cpc',
-                r"C:\My Photos\2017-05-13 GBR\1.jpg", [
-                    (49*15, 49*15, 'A'),
-                    (59*15, 39*15, 'B')]),
-            self.make_cpc_file(
-                'GBR_2.cpc',
-                r"C:\My Photos\2017-05-13 GBR\2.jpg", [
-                    (69*15, 29*15, 'A'),
-                    (79*15, 19*15, 'A')]),
-        ]
-        # Save the file content for comparison purposes.
-        img1_expected_cpc_content = cpc_files[0].read()
-        img2_expected_cpc_content = cpc_files[1].read()
-        # Reset the file pointer so that the views can read from the start.
-        cpc_files[0].seek(0)
-        cpc_files[1].seek(0)
-
-        self.preview_cpc_annotations(self.user, self.source, cpc_files)
-        self.upload_annotations(self.user, self.source)
-
-        self.img1.refresh_from_db()
-        self.assertEqual(self.img1.cpc_content, img1_expected_cpc_content)
-        self.assertEqual(self.img1.cpc_filename, 'GBR_1.cpc')
-        self.img2.refresh_from_db()
-        self.assertEqual(self.img2.cpc_content, img2_expected_cpc_content)
-        self.assertEqual(self.img2.cpc_filename, 'GBR_2.cpc')
-
-    def test_source_fields(self):
-        cpc_files = [
-            self.make_cpc_file(
-                '1.cpc',
-                r"C:\My Photos\2017-05-13 GBR\1.jpg", [
-                    (49*15, 49*15, 'A'),
-                    (59*15, 39*15, 'B')],
-                codes_filepath=r'C:\PROGRA~4\CPCE_4~1\SHALLO~1.TXT'),
-            self.make_cpc_file(
-                '2.cpc',
-                r"C:\My Photos\2017-05-13 GBR\2.jpg", [
-                    (69*15, 29*15, 'A'),
-                    (79*15, 19*15, 'A')],
-                codes_filepath=r'C:\My Photos\CPCe codefiles\GBR codes.txt'),
-        ]
-        self.preview_cpc_annotations(self.user, self.source, cpc_files)
-        self.upload_annotations(self.user, self.source)
-
-        self.source.refresh_from_db()
-        # Although it's an implementation detail and not part of spec,
-        # the last uploaded CPC should have its values used in a multi-CPC
-        # upload.
-        self.assertEqual(
-            self.source.cpce_code_filepath,
-            r'C:\My Photos\CPCe codefiles\GBR codes.txt')
-        self.assertEqual(
-            self.source.cpce_image_dir, r'C:\My Photos\2017-05-13 GBR')
